Got the second table working

+++ src/main/scala/com/snowplowanalytics/snowpik/SnowPikApp.scala
@@ -72,7 +72,7 @@ object SnowPikApp {
                                            "Flags that the generated .csv files should not be uploaded to S3")
 
   // Optional site id separator
-  val siteId = parser.option[Int](List("s", "site"), "id",
+  val siteId = parser.option[Long](List("s", "site"), "id",
                                       "Piwik site id to extract data for. Defaults to 1")
 
   // Optional time period
@@ -102,7 +102,7 @@ object SnowPikApp {
       SnowPik(config = config.value.getOrElse(ConfigFactory.load("default")), // Fall back to the /resources/default.conf
               period = period.value.getOrElse(TimePeriod.YESTERDAY),          // Default to yesterday's data
               upload = !(noUpload.value.getOrElse(false))                     // Default to true (i.e. we upload to S3)
-      ).run(siteId = siteId.value.getOrElse(1))                               // Default to 1 for the site ID
+      ).run(siteId = siteId.value.getOrElse(1L))                               // Default to 1 for the site ID
     } catch {
       case e: ArgotUsageException => println(e.message)
     }
