+++ src/main/scala/com/snowplowanalytics/snowpik/models/LogVisit.scala
@@ -273,7 +273,7 @@ class ExtractableLogVisit(table: Table[LogVisit]) extends Extractor[CsvFile] {
   /**
    * Exports this table to .csv
    */
-  def ~>(logFile: CsvFile)(implicit siteId: Int, folder: String) {
+  def ~>(logFile: CsvFile)(implicit siteId: Long, folder: String) {
 
     inTransaction {
       from (table)(t =>
